pyproject.toml
[project]
name = "mssbot-fluxer"
version = "0.1.0"
description = "MSSBot but for fluxer."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fluxer-py>=0.4.2",
]
